fix: fix logical errors

+++ controllers/gamecontroller.js
@@ -6,12 +6,13 @@ router.get('/all', (req, res) => {
         .then(
             function findSuccess(data) {
                 res.status(200).json({
-                    games: games,
+                    games: data,
                     message: "Data fetched."
                 })
             },
 
-            function findFail() {
+            function findFail(data) {
+                console.log(data)
                 res.status(500).json({
                     message: "Data not found"
                 })
@@ -39,7 +40,7 @@ router.get('/:id', (req, res) => {
 router.post('/create', (req, res) => {
     Game.create({
         title: req.body.game.title,
-        owner_id: req.body.user.id,
+        owner_id: req.user.id,
         studio: req.body.game.studio,
         esrb_rating: req.body.game.esrb_rating,
         user_rating: req.body.game.user_rating,
@@ -70,47 +71,62 @@ router.put('/update/:id', (req, res) => {
         {
             where: {
                 id: req.params.id,
-                owner_id: req.user
-            }
+                owner_id: req.user.id
+            },
         })
         .then(
-            function updateSuccess(game) {
-                res.status(200).json({
-                    game: game,
-                    message: "Successfully updated."
-                })
+            function resHandler(response) {
+                if(response[0] === 0) {
+                    res.status(404).json({
+                        message: "Data not found."
+                    })
+                } else {
+                    res.status(200).json({
+                        message: "Successfully updated."
+                    })
+                }
             },
-
             function updateFail(err) {
                 res.status(500).json({
                     message: err.message
                 })
             }
-
         )
 })
 
 router.delete('/remove/:id', (req, res) => {
-    Game.destroy({
-        where: {
-            id: req.params.id,
-            owner_id: req.user.id
-        }
-    })
+    return Game.findOne({ where: { id: req.params.id, owner_id: req.user.id } })
     .then(
-        function deleteSuccess(game) {
-            res.status(200).json({
-                game: game,
-                message: "Successfully deleted"
-            })
+        function resHandler(gameModel) {
+            if(gameModel === null) {
+                res.status(404).json({
+                    message: "Data not found."
+                })
+            } else {
+                Game.destroy({
+                    where: {
+                        id: req.params.id,
+                        owner_id: req.user.id
+                    }
+                })
+                .then(
+                 function successDelete() {
+                    res.status(200).json({
+                        message: "Successfully deleted"
+                    })
+                },
+                function deleteFail(err) {
+                    res.status(500).json({
+                        error: err.message
+                    })
+                })    
+            }
         },
-
         function deleteFail(err) {
             res.status(500).json({
                 error: err.message
             })
-        }
-    )
-})
+        })
+})    
 
 module.exports = router;
